backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<0.31.0
SQLAlchemy>=2.0.28,<3.0.0
psycopg[binary]>=3.1.18,<4.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
pydantic-settings>=2.2.1,<3.0.0
email-validator>=2.1.0,<3.0.0
alembic>=1.12.1,<2.0.0
python-dotenv>=1.0.0,<2.0.0
